backend/requirements.txt
Django>=5.0,<6
psycopg2-binary>=2.9
djangorestframework>=3.14
django-cors-headers>=4.3
